refactor(bridge): share backward line scanner between codex and pi history
Extract the identical 64k backward-chunk tail scan into backward-line-scan.js and switch both pagers to it, keeping each harness's cursor format and stop semantics. OMP untouched (forward-stream reader, different technique).

Tests: codex/pi/omp history suites 27/27 pass; full bridge suite unchanged (1 pre-existing Windows symlink EPERM in project-catalog).
